move converter and formatter given as example to test case
