+++ python_jsonschema_objects/validators.py
@@ -15,7 +15,6 @@ SCHEMA_TYPE_MAPPING = (
 )
 """Sequence of schema type mappings to be checked in precedence order."""
 
-
 class ValidationError(Exception):
     pass
 
@@ -169,28 +168,4 @@ def check_type(param, value, type_data):
 
 converter_registry = ValidatorRegistry()
 
-@converter_registry.register(name='boolean')
-def convert_boolean(param, value, _):
-    if isinstance(value, six.string_types):
-        vl = value.lower()
-        if vl in ['true', 'yes', 'ok']:
-            return True
-        if vl in ['false', 'no', 'wrong']:
-            return False
-    return value
-
-
 formatter_registry = ValidatorRegistry()
-
-@formatter_registry.register(name='number')
-def format_number(param, value, details):
-    if 'format' in details:
-        frmt = details['format']
-        try:
-            if '{' in frmt:
-                return frmt.format(value)
-            if '%' in frmt:
-                return frmt % value
-        except ValueError as er:
-            pass
-    return value
